Add 1D implementation of a fast LCT algorithm. (WIP)
This remains a work in progress.
The essential pieces are all in place,
but they still require testing.

I am currently working on the resampling algorithm,
which must be executed before a Chirp Multiplication.
A principal difficulty is meeting a set of conflicting
demands: keep the resampled points within the domain
spanned by the original points, maintain the parity
of the number of points, and keep 0 as one of the
points. I think I have to ditch that last one.